Move `DepNodeFilter` and `EdgeFilter`.
They're currently in `rustc_middle::dep_graph::debug` but they're only
used in `rustc_incremental::assert_dep_graph`. This commit moves them to
that module.

This avoids some cross-crate dependencies (and `pub` visibilities), and
makes `rustc_middle` a little smaller, both of which are good.
